refactor : 수동/자동 로또 번호 합치기 수정
