+++ src/main/java/lotto/factory/LottosFactory.java
@@ -1,0 +1,10 @@
+package lotto.factory;
+
+import lotto.model.Lottos;
+
+public abstract class LottosFactory {
+
+    public abstract Lottos generate(int numberOfLotto);
+
+
+}
